Creates pokemon class
